converter/requirements.txt
pydicom>=2.4
Pillow>=10.0
numpy>=1.24